% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amolim.R
\docType{data}
\name{amolim}
\alias{amolim}
\title{Censo Experimental de Limeira de 1988}
\format{Esse data frame contém as seguintes colunas:
 \describe{
 \item{setor}{Identificação do setor censitário}
 \item{np}{Número de domicílios no setor censitário}
 \item{domic}{Identificação do domicílio}
 \item{renda}{Renda mensal (em reais)}
 \item{lrenda}{Logaritmo da renda mensal}
 \item{raca}{Cor/Raça (\code{1}=Branca e \code{2}=Outras)}
 \item{estudo}{Anos de estudo (\code{1}=Até 4 anos; \code{2}=De 5 a 8 anos e
 \code{3}=Mais de 8 anos)}
 \item{idade}{Idade em anos completos (\code{1}=Até 25 anos; \code{2}=De 26 a 40 anos;
 \code{3}=De 41 a 60 anos e \code{4} =Mais de 60 anos)}
\item{na}{Número de pessoas na amostra do setor censitário}
\item{peso}{Peso do domicílio no setor (2º estágio de seleção)}
\item{domtot}{Número total de domicílios no município}
\item{peso1}{Peso do setor censitário (1º estágio de seleção)}
\item{pesof}{Peso final do domicílio}
 }}
\source{
Microdados da Amostra do Censo Experimental de Limeira de 1988.
}
\usage{
amolim
}
\description{
Dados da amostra do Censo Experimental de Limeira de 1988.
}
\details{
O data frame \code{amolim} contém 706 linhas e 14 colunas
}
\examples{
amolim
}
\references{
Pessoa, D.G.C. e Silva, P.L.N. (1998). Análise de Dados Amostrais Complexos, 13° SINAPE, Associção Brasileira de Estatística.
}
\keyword{datasets}
