% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnadrj90.R
\docType{data}
\name{pnadrj90}
\alias{pnadrj90}
\title{Dados da Pesquisa Nacional por Amostra de Domicílios de 1990 - RJ}
\format{Esse data frame contém as seguintes colunas:
 \describe{
 \item{stra}{estratos}
 \item{psu}{unidade primária de amostragem}
 \item{pesopes}{pesos}
 \item{informal}{trabalho informal}
 \item{sx}{sexo}
 \item{id}{idade}
 \item{ae}{anos de estudo}
 \item{ht}{horas de trabalho por dia}
 \item{re}{salário}
 \item{um}{valores 1}
 }}
\usage{
pnadrj90
}
\description{
O data frame  \code{pnadrj90} tem 6507 linha e 10 colunas
}
\examples{
pnadrj90
}
\keyword{datasets}
