% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popul.R
\docType{data}
\name{popul}
\alias{popul}
\title{Dados de uma população de 749 empresas}
\format{Esse data frame contém as seguintes colunas:
 \describe{
 \item{id}{código de identificação}
 \item{po}{número de pessoas ocupadas na empresa}
 \item{sal}{total de salários pagos na empresa}
 \item{rec}{receita total da empresa}
 \item{estrat}{código do estrato}
 }}
\usage{
popul
}
\description{
O data frame \code{popul} tem  749 linhas e 5 colunas.
}
\examples{
popul
}
\keyword{datasets}
