% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv.R
\docType{data}
\name{ppv}
\alias{ppv}
\title{Dados da Pesquisa de Padrões de Vida(PPV)}
\format{Um data frame com 53940 linhas and 10 variáveis:
\describe{
  \item{serie}{código de identificação}
  \item{ident}{código de identificação}
  \item{codmor}{código de identificação do morador}
  \item{v04a01}{1- pode ler; 2- não apto a ler}
  \item{v04a02}{1- abilidade de leitura; 2- nenhuma abilidade de #'  leitura}
  \item{v04a03}{relacionado a capacidade de leitura}
  \item{estratof}{estrato final}
  \item{peso1}{pesos no primeiro estágio}
  \item{peso2}{pesos no segundo estágio}
  \item{pesof}{pesos finais}
  \item{nsetor}{identificação da unidade secundária de amostragem}
  \item{regiao}{1- Nordeste; 2- Sudeste}
  \item{v02a08}{idades}
  }}
\source{
Pesquisa de Padrão de Vida do IBGE
}
\usage{
ppv
}
\description{
O data frame \code{ppv} tem 19409 linhas e 13 colunas
}
\details{
Se \code{v04a01} é 2 ou \code{v04a02} é 2 então a pessoa é analfabeta.
}
\examples{
ppv
}
\references{
Pessoa & Silva, Análise de Dados Amostrais Complexos
}
\keyword{datasets}
